refactor(projects): group in directory & split models
