+++ src/app/projects/credit.ts
@@ -1,0 +1,4 @@
+export interface Credit {
+  readonly role: string
+  readonly authorSlug: string
+}
