+++ src/app/projects/lookbook-name-and-slug.ts
@@ -1,0 +1,4 @@
+export interface LookbookNameAndSlug {
+  readonly slug: string
+  readonly name: string
+}
