+++ src/app/projects/project-list-item.ts
@@ -1,0 +1,9 @@
+import { ImageAsset } from '../common/images/image-asset'
+import { Project } from './project'
+
+export type ProjectListItem = Project & ListItemExtraData
+
+export interface ListItemExtraData {
+  readonly previewImages?: ReadonlyArray<ImageAsset>
+  readonly hasContent: boolean
+}
