+++ src/app/projects/project-page/project-page.component.scss
