+++ src/app/projects/project-page/project-page.component.spec.ts
@@ -2,9 +2,9 @@ import { ComponentFixture, TestBed } from '@angular/core/testing'
 
 import { ProjectPageComponent } from './project-page.component'
 import { MockComponents, MockProvider } from 'ng-mocks'
-import { ProjectsService } from '../projects-page/projects.service'
-import { Project } from '../projects-page/project-item/project-item'
-import { ImagesSwiperComponent } from '../images-swiper/images-swiper.component'
+import { ProjectsService } from '../projects.service'
+import { Project } from '../project-list-item'
+import { ImagesSwiperComponent } from '../../images-swiper/images-swiper.component'
 import { ProjectLookbooksService } from './project-lookbooks.service'
 import { ProjectImagesService } from './project-images.service'
 import { of } from 'rxjs'
