+++ src/app/projects/projects-page/project-list-item/project-list-item.component.scss
