+++ src/app/projects/projects-page/projects-page.component.html
@@ -1,6 +1,6 @@
-<app-project-item
+<app-project-list-item
   *ngFor="let item of projects | async; index as index"
   [item]="item"
   [priority]="index < MAX_PROJECTS_PER_PAGE"
   [class.has-preview-images]="item.previewImages && item.previewImages.length"
-></app-project-item>
+></app-project-list-item>
